Reformat date, time, and datetime columns

Reformat date, time, and datetime columns

clean_date_columns(x)

Arguments

x
tibble/df; dataset with the columns to be formatted

Value

tibble/df; dataset with reformatted columns (if any needed reformatting)

Examples

generation_dataset_unclean <- as.data.frame(
apply(generation_dataset_example, 2, as.character)
) #Create a version of the example generation dataset with character columns
clean_date_columns(generation_dataset_unclean)

#>    record_type settlement_date settlement_period           spot_time  ccgt oil
#> 1     FUELINST      2019-07-01                 2 2019-07-01 00:00:00  4101   0
#> 2     FUELINST      2019-07-01                 3 2019-07-01 00:05:00  4354   0
#> 3     FUELINST      2019-07-01                 3 2019-07-01 00:10:00  4388   0
#> 4     FUELINST      2019-07-01                 3 2019-07-01 00:15:00  4290   0
#> 5     FUELINST      2019-07-01                 3 2019-07-01 00:20:00  4237   0
#> 6     FUELINST      2019-07-01                 3 2019-07-01 00:25:00  4176   0
#> 7     FUELINST      2019-07-01                 3 2019-07-01 00:30:00  4074   0
#> 8     FUELINST      2019-07-01                 4 2019-07-01 00:35:00  4160   0
#> 9     FUELINST      2019-07-01                 4 2019-07-01 00:40:00  4115   0
#> 10    FUELINST      2019-07-01                 4 2019-07-01 00:45:00  4107   0
#> 11    FUELINST      2019-07-01                 4 2019-07-01 00:50:00  4104   0
#> 12    FUELINST      2019-07-01                 4 2019-07-01 00:55:00  4088   0
#> 13    FUELINST      2019-07-01                 4 2019-07-01 01:00:00  4103   0
#> 14    FUELINST      2019-07-01                 5 2019-07-01 01:05:00  4185   0
#> 15    FUELINST      2019-07-01                 5 2019-07-01 01:10:00  4123   0
#> 16    FUELINST      2019-07-01                 5 2019-07-01 01:15:00  4189   0
#> 17    FUELINST      2019-07-01                 5 2019-07-01 01:20:00  4232   0
#> 18    FUELINST      2019-07-01                 5 2019-07-01 01:25:00  4354   0
#> 19    FUELINST      2019-07-01                 5 2019-07-01 01:30:00  4390   0
#> 20    FUELINST      2019-07-01                 6 2019-07-01 01:35:00  4336   0
#> 21    FUELINST      2019-07-01                 6 2019-07-01 01:40:00  4186   0
#> 22    FUELINST      2019-07-01                 6 2019-07-01 01:45:00  4102   0
#> 23    FUELINST      2019-07-01                 6 2019-07-01 01:50:00  4181   0
#> 24    FUELINST      2019-07-01                 6 2019-07-01 01:55:00  4160   0
#> 25    FUELINST      2019-07-01                 6 2019-07-01 02:00:00  4120   0
#> 26    FUELINST      2019-07-01                 7 2019-07-01 02:05:00  4174   0
#> 27    FUELINST      2019-07-01                 7 2019-07-01 02:10:00  4128   0
#> 28    FUELINST      2019-07-01                 7 2019-07-01 02:15:00  4067   0
#> 29    FUELINST      2019-07-01                 7 2019-07-01 02:20:00  4133   0
#> 30    FUELINST      2019-07-01                 7 2019-07-01 02:25:00  4173   0
#> 31    FUELINST      2019-07-01                 7 2019-07-01 02:30:00  4169   0
#> 32    FUELINST      2019-07-01                 8 2019-07-01 02:35:00  4156   0
#> 33    FUELINST      2019-07-01                 8 2019-07-01 02:40:00  4157   0
#> 34    FUELINST      2019-07-01                 8 2019-07-01 02:45:00  4153   0
#> 35    FUELINST      2019-07-01                 8 2019-07-01 02:50:00  4096   0
#> 36    FUELINST      2019-07-01                 8 2019-07-01 02:55:00  4004   0
#> 37    FUELINST      2019-07-01                 8 2019-07-01 03:00:00  3994   0
#> 38    FUELINST      2019-07-01                 9 2019-07-01 03:05:00  4038   0
#> 39    FUELINST      2019-07-01                 9 2019-07-01 03:10:00  4048   0
#> 40    FUELINST      2019-07-01                 9 2019-07-01 03:15:00  4161   0
#> 41    FUELINST      2019-07-01                 9 2019-07-01 03:20:00  4183   0
#> 42    FUELINST      2019-07-01                 9 2019-07-01 03:25:00  4177   0
#> 43    FUELINST      2019-07-01                 9 2019-07-01 03:30:00  4126   0
#> 44    FUELINST      2019-07-01                10 2019-07-01 03:35:00  4127   0
#> 45    FUELINST      2019-07-01                10 2019-07-01 03:40:00  4165   0
#> 46    FUELINST      2019-07-01                10 2019-07-01 03:45:00  4108   0
#> 47    FUELINST      2019-07-01                10 2019-07-01 03:50:00  4127   0
#> 48    FUELINST      2019-07-01                10 2019-07-01 03:55:00  4186   0
#> 49    FUELINST      2019-07-01                10 2019-07-01 04:00:00  4237   0
#> 50    FUELINST      2019-07-01                11 2019-07-01 04:05:00  4357   0
#> 51    FUELINST      2019-07-01                11 2019-07-01 04:10:00  4301   0
#> 52    FUELINST      2019-07-01                11 2019-07-01 04:15:00  4397   0
#>    coal nuclear  wind   ps npshyd ocgt other intfr intirl intned intew biomass
#> 1     0    4673  8782    0     85    4    79   327    117    228   379    1178
#> 2     0    4668  8747    0     85    4    79    44    117      0   405    1207
#> 3     0    4656  8794    0     85    5    79     0    117      0   430    1199
#> 4     0    4656  8875    0     85    3    81     0    117      0   454    1200
#> 5     0    4665  8880    0     85    3    80     0    117      0   476    1247
#> 6     0    4661  8968    0     85    3    80     0    117      0   482    1243
#> 7     0    4660  9011    0     85    5    80     0    117      0   482    1234
#> 8     0    4662  9071    0     85    4    80     0    117      0   471    1252
#> 9     0    4665  9186    0     85    3    81     0    117      0   446    1247
#> 10    0    4673  9153    0     85    2    81     0    117      0   428    1257
#> 11    0    4672  9034    0     85    3    82     0    117      0   428    1252
#> 12    0    4671  9002    0     85    3    81     0    117      0   428    1282
#> 13    0    4673  8913    0     85    3    81    16    117      0   440    1298
#> 14    0    4666  8921    0     85    3    81   478    117      0   466    1287
#> 15    0    4666  9189    0     85    3    80   629    117      0   473    1278
#> 16    0    4668  9253    0     85    3    80   630    117      0   473    1283
#> 17    0    4677  9265    0     85    3    80   629    117      0   473    1282
#> 18    0    4676  9228    0     85    3    81   629    117      0   473    1234
#> 19    0    4671  9181    0     85    3    80   629    117      0   467    1230
#> 20    0    4676  9094    0     85    3    80   629    117      0   443    1312
#> 21    0    4677  9044    0     85    4    79   630    117      0   437    1381
#> 22    0    4676  9037    0     85    2    79   629    117      0   437    1405
#> 23    0    4669  9029    0     85    4    79   603    117      0   437    1410
#> 24    0    4675  8941    0     85    2    79   575    117      0   437    1409
#> 25    0    4674  8941    0     84    3    79   575    117      0   429    1400
#> 26    0    4678  8966    0     84    2    78   548    117      0   404    1415
#> 27    0    4668  9043    0     84    4    78   544    117      0   387    1439
#> 28    0    4676  9058    0     84    5    78   544    117      0   387    1468
#> 29    0    4681  8932    0     83    3    78   544    117      0   387    1486
#> 30    0    4681  8867    0     82    4    78   544    117      0   389    1489
#> 31    0    4680  8837    0     82    4    78   544    117      0   394    1482
#> 32    0    4679  8872    0     82    3    78   546    117      0   415    1487
#> 33    0    4671  8917    0     82    3    77   546    117      0   423    1485
#> 34    0    4668  8932    0     82    3    77   546    117      0   423    1486
#> 35    0    4670  8958    0     82    3    77   546    117      0   423    1487
#> 36    0    4677  9090    0     82    4    77   546    117      0   422    1510
#> 37    0    4677  9275    0     82    3    77   539    117      0   409    1563
#> 38    0    4681  9347    0     82    4    77   426    117      0   383    1586
#> 39    0    4678  9472    0     82    3    78   413    117      0   365    1575
#> 40    0    4677  9540    0     82    2    77   413    117      0   365    1587
#> 41    0    4673  9458    0     82    5    77   413    117      0   365    1587
#> 42    0    4669  9385    0     82    3    77   413    117      0   366    1583
#> 43    0    4669  9423    0     82    3    77   413    117      0   368    1572
#> 44    0    4674  9398    0     82    3    77   435    117      0   349    1576
#> 45    0    4681  9352    0     82    3    77   436    117      0   342    1584
#> 46    0    4678  9366    0     82    4    77   436    117      0   342    1572
#> 47    0    4678  9375    0     82    2    77   436    117      0   342    1499
#> 48    0    4678  9333    0     82    4    77   436    117      0   347    1446
#> 49    0    4677  9289    0     87    3    77   332    117      0   367    1459
#> 50    0    4671  9034    0    101    3    77     0    107      0   392    1480
#> 51    0    4667  8872    0    102    2    77     0    108      0   415    1509
#> 52    0    4676  8883    0    101    2    77     0    108      0   419    1558
#>    intnem
#> 1       0
#> 2       0
#> 3       0
#> 4       0
#> 5       0
#> 6       0
#> 7       0
#> 8       0
#> 9       0
#> 10      0
#> 11      0
#> 12      0
#> 13      0
#> 14      0
#> 15      0
#> 16      0
#> 17      0
#> 18      0
#> 19      0
#> 20      0
#> 21      0
#> 22      0
#> 23      0
#> 24      0
#> 25      0
#> 26      0
#> 27      0
#> 28      0
#> 29      0
#> 30      0
#> 31      0
#> 32      0
#> 33      0
#> 34      0
#> 35      0
#> 36      0
#> 37      0
#> 38      0
#> 39      0
#> 40      0
#> 41      0
#> 42      0
#> 43      0
#> 44      0
#> 45      0
#> 46      0
#> 47      0
#> 48      0
#> 49      0
#> 50      0
#> 51      0
#> 52      0
#>  [ reached 'max' / getOption("max.print") -- omitted 525 rows ]